#3 Horizontal lines do now consider folded areas.
There are now new instances of LinePaintingPaintListener created,
instead of one static one. That is because the painters must know about
the ITextViewer which they are operating on to get the information about
folded areas. Unfortunately we can't get from the StyledText back "up"
to the ITextViewer, so we have to save the instance we are operating on.